+++ Products/CMFPlone/interfaces/syndication.py
@@ -1,6 +1,7 @@
 from zope.interface import Interface
 from zope import schema
 from Products.CMFPlone import PloneMessageFactory as _
+from Products.CMFPlone.validators import positiveNumber
 
 
 class ISyndicatable(Interface):
@@ -174,7 +175,8 @@ class ISiteSyndicationSettings(Interface):
         title=_(u'label_syndication_max_items',
                 default=u'Maximum items'),
         description=_(u'help_syndication_max_items',
-                default=u'Maximum number of items that will be syndicated.'),
+                      default=u'Maximum number of items that will be syndicated.'),
+        constraint=positiveNumber,
         default=15)
 
     allowed_feed_types = schema.Tuple(
@@ -228,16 +230,18 @@ class IFeedSettings(Interface):
 
     render_body = schema.Bool(
         title=_(u'Render Body'),
-        description=_(u'help_render_body',
-                default=u'If body text available for item, '
-                        u'render it, otherwise use description.'),
+        description=_(
+            u'help_render_body',
+            default=u'If body text available for item, '
+                    u'render it, otherwise use description.'),
         default=False)
 
     max_items = schema.Int(
         title=_(u'label_syndication_max_items',
                 default=u'Maximum items'),
-        description=_(u'help_syndication_max_items',
-                default=u'Maximum number of items that will be syndicated.'),
+        description=_(
+            u'help_syndication_max_items',
+            default=u'Maximum number of items that will be syndicated.'),
         default=15)